WIP: Implement mapping for defendant parent/guardian details in fines MAC payload
