+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/fines-mac-payload-map-account/fines-mac-payload-map-account-defendant-company.utils.spec.ts
@@ -1,0 +1,1 @@
+describe('mapAccountDefendantCompanyPayload', () => {});
