+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/fines-mac-payload-map-account/fines-mac-payload-map-account-defendant-parent-guardian.utils.ts
@@ -1,69 +1,114 @@
-import { IFinesMacCompanyDetailsAliasState } from '../../../../fines-mac-company-details/interfaces/fines-mac-company-details-alias-state.interface';
 import { IFinesMacState } from '../../../../interfaces/fines-mac-state.interface';
 import { IFinesMacAddAccountPayload } from '../../interfaces/fines-mac-payload-add-account.interfaces';
-import { IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete } from '../interfaces/fines-mac-payload-account-defendant-debtor-detail-alias-complete.interface';
 
-const mapAccountDefendantCompanyDebtorDetailsAliases = (
-  payloadAccountDefendantCompanyDebtorDetailsAliases:
-    | IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete[]
-    | null,
-): IFinesMacCompanyDetailsAliasState[] => {
-  return payloadAccountDefendantCompanyDebtorDetailsAliases
-    ? payloadAccountDefendantCompanyDebtorDetailsAliases.map((alias, index) => {
-        return {
-          [`fm_company_details_alias_organisation_name_${index}`]: alias.alias_company_name,
-        };
-      })
-    : [];
-};
-
-const mapAccountDefendantCompanyDebtorDetails = (
+const mapAccountDefendantParentGuardianDetailsPayloadDebtorDetailsToFinesMacState = (
   finesMacState: IFinesMacState,
   payload: IFinesMacAddAccountPayload,
 ): IFinesMacState => {
-  const payloadAccountDefendantDebtorDetails = payload.account.defendant.debtor_detail;
+  const payloadAccountDefendantParentGuardianDebtorDetails = payload.account.defendant.parent_guardian?.debtor_detail;
 
-  const aliases = payloadAccountDefendantDebtorDetails?.aliases
-    ? mapAccountDefendantCompanyDebtorDetailsAliases(payloadAccountDefendantDebtorDetails?.aliases)
-    : finesMacState.companyDetails.formData.fm_company_details_aliases;
+  const aliases = payloadAccountDefendantParentGuardianDebtorDetails?.aliases
+    ? payloadAccountDefendantParentGuardianDebtorDetails.aliases.map((alias, index) => {
+        return {
+          [`fm_parent_guardian_details_alias_forenames_${index}`]: alias.alias_forenames,
+          [`fm_parent_guardian_details_alias_surname_${index}`]: alias.alias_surname,
+        };
+      })
+    : finesMacState.parentGuardianDetails.formData.fm_parent_guardian_details_aliases;
 
-  finesMacState.companyDetails.formData = {
-    ...finesMacState.companyDetails.formData,
-    fm_company_details_aliases: aliases,
+  finesMacState.parentGuardianDetails.formData = {
+    ...finesMacState.parentGuardianDetails.formData,
+    fm_parent_guardian_details_vehicle_make: payloadAccountDefendantParentGuardianDebtorDetails?.vehicle_make ?? null,
+    fm_parent_guardian_details_vehicle_registration_mark:
+      payloadAccountDefendantParentGuardianDebtorDetails?.vehicle_registration_mark ?? null,
+    fm_parent_guardian_details_aliases: aliases,
+  };
+
+  finesMacState.employerDetails.formData = {
+    ...finesMacState.employerDetails.formData,
+    fm_employer_details_employer_reference:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employee_reference ?? null,
+    fm_employer_details_employer_company_name:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_company_name ?? null,
+    fm_employer_details_employer_address_line_1:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_address_line_1 ?? null,
+    fm_employer_details_employer_address_line_2:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_address_line_2 ?? null,
+    fm_employer_details_employer_address_line_3:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_address_line_3 ?? null,
+    fm_employer_details_employer_address_line_4:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_address_line_4 ?? null,
+    fm_employer_details_employer_address_line_5:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_address_line_5 ?? null,
+    fm_employer_details_employer_post_code:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_post_code ?? null,
+    fm_employer_details_employer_telephone_number:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_telephone_number ?? null,
+    fm_employer_details_employer_email_address:
+      payloadAccountDefendantParentGuardianDebtorDetails?.employer_email_address ?? null,
   };
 
   finesMacState.languagePreferences.formData = {
     ...finesMacState.languagePreferences.formData,
-    fm_language_preferences_document_language: payloadAccountDefendantDebtorDetails?.document_language ?? null,
-    fm_language_preferences_hearing_language: payloadAccountDefendantDebtorDetails?.hearing_language ?? null,
+    fm_language_preferences_document_language:
+      payloadAccountDefendantParentGuardianDebtorDetails?.document_language ?? null,
+    fm_language_preferences_hearing_language:
+      payloadAccountDefendantParentGuardianDebtorDetails?.hearing_language ?? null,
   };
 
   return finesMacState;
 };
 
-export const mapAccountDefendantCompanyPayload = (
+const mapAccountDefendantParentGuardianDetailsPayloadToFinesMacState = (
+  finesMacState: IFinesMacState,
+  payload: IFinesMacAddAccountPayload,
+): IFinesMacState => {
+  const payloadAccountDefendantParentGuardian = payload.account.defendant.parent_guardian;
+
+  finesMacState.parentGuardianDetails.formData = {
+    ...finesMacState.parentGuardianDetails.formData,
+    fm_parent_guardian_details_surname: payloadAccountDefendantParentGuardian?.surname ?? null,
+    fm_parent_guardian_details_forenames: payloadAccountDefendantParentGuardian?.forenames ?? null,
+    fm_parent_guardian_details_dob: payloadAccountDefendantParentGuardian?.dob ?? null,
+    fm_parent_guardian_details_national_insurance_number:
+      payloadAccountDefendantParentGuardian?.national_insurance_number ?? null,
+    fm_parent_guardian_details_address_line_1: payloadAccountDefendantParentGuardian?.address_line_1 ?? null,
+    fm_parent_guardian_details_address_line_2: payloadAccountDefendantParentGuardian?.address_line_2 ?? null,
+    fm_parent_guardian_details_address_line_3: payloadAccountDefendantParentGuardian?.address_line_3 ?? null,
+    fm_parent_guardian_details_post_code: payloadAccountDefendantParentGuardian?.post_code ?? null,
+  };
+
+  finesMacState.contactDetails.formData = {
+    ...finesMacState.contactDetails.formData,
+    fm_contact_details_telephone_number_home: payloadAccountDefendantParentGuardian?.telephone_number_home ?? null,
+    fm_contact_details_telephone_number_business:
+      payloadAccountDefendantParentGuardian?.telephone_number_business ?? null,
+    fm_contact_details_telephone_number_mobile: payloadAccountDefendantParentGuardian?.telephone_number_mobile ?? null,
+    fm_contact_details_email_address_1: payloadAccountDefendantParentGuardian?.email_address_1 ?? null,
+    fm_contact_details_email_address_2: payloadAccountDefendantParentGuardian?.email_address_2 ?? null,
+  };
+
+  return mapAccountDefendantParentGuardianDetailsPayloadDebtorDetailsToFinesMacState(finesMacState, payload);
+};
+
+export const mapAccountDefendantParentGuardianPayload = (
   finesMacState: IFinesMacState,
   payload: IFinesMacAddAccountPayload,
 ): IFinesMacState => {
   const payloadAccountDefendant = payload.account.defendant;
 
-  finesMacState.companyDetails.formData = {
-    ...finesMacState.companyDetails.formData,
-    fm_company_details_organisation_name: payloadAccountDefendant.organisation_name,
-    fm_company_details_address_line_1: payloadAccountDefendant.address_line_1,
-    fm_company_details_address_line_2: payloadAccountDefendant.address_line_2,
-    fm_company_details_address_line_3: payloadAccountDefendant.address_line_3,
-    fm_company_details_postcode: payloadAccountDefendant.post_code,
+  finesMacState.personalDetails.formData = {
+    ...finesMacState.personalDetails.formData,
+    fm_personal_details_title: payloadAccountDefendant.title,
+    fm_personal_details_surname: payloadAccountDefendant.surname,
+    fm_personal_details_forenames: payloadAccountDefendant.forenames,
+    fm_personal_details_dob: payloadAccountDefendant.dob,
+    fm_personal_details_address_line_1: payloadAccountDefendant.address_line_1,
+    fm_personal_details_address_line_2: payloadAccountDefendant.address_line_2,
+    fm_personal_details_address_line_3: payloadAccountDefendant.address_line_3,
+    fm_personal_details_post_code: payloadAccountDefendant.post_code,
+    fm_personal_details_national_insurance_number: payloadAccountDefendant.national_insurance_number,
   };
 
-  finesMacState.contactDetails.formData = {
-    ...finesMacState.contactDetails.formData,
-    fm_contact_details_telephone_number_home: payloadAccountDefendant.telephone_number_home,
-    fm_contact_details_telephone_number_business: payloadAccountDefendant.telephone_number_business,
-    fm_contact_details_telephone_number_mobile: payloadAccountDefendant.telephone_number_mobile,
-    fm_contact_details_email_address_1: payloadAccountDefendant.email_address_1,
-    fm_contact_details_email_address_2: payloadAccountDefendant.email_address_2,
-  };
-
-  return mapAccountDefendantCompanyDebtorDetails(finesMacState, payload);
+  return mapAccountDefendantParentGuardianDetailsPayloadToFinesMacState(finesMacState, payload);
 };
